# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/logistics/product/{id}": {
            "get": {
                "description": "读取货物批次详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "物流企业-Logistics"
                ],
                "summary": "读取货物批次详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "批次编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.ProductInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/logistics/tracking/{id}": {
            "get": {
                "description": "读取物流详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "物流企业-Logistics"
                ],
                "summary": "读取物流详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "物流号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.LogisticsCombinedRecord"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/logistics/update/{id}": {
            "post": {
                "description": "物流记录上链",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "物流企业-Logistics"
                ],
                "summary": "物流记录上链",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateLogisticRecordInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "物流追踪ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/regulator/private/{id}": {
            "get": {
                "description": "读取物流详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "监管机构-Regulator"
                ],
                "summary": "读取物流详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "物流号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.LogisticsCombinedRecord"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/regulator/product/{id}": {
            "get": {
                "description": "读取货物批次详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "监管机构-Regulator"
                ],
                "summary": "读取货物批次详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "批次编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.ProductInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/regulator/tracking/{id}": {
            "get": {
                "description": "读取物流详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "监管机构-Regulator"
                ],
                "summary": "读取物流详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "物流号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.LogisticsRecord"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dao.LogisticsCombinedRecord": {
            "type": "object",
            "properties": {
                "private": {
                    "$ref": "#/definitions/dao.PrivateLogisticsRecord"
                },
                "public": {
                    "$ref": "#/definitions/dao.LogisticsRecord"
                }
            }
        },
        "dao.LogisticsRecord": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.RecordSubItem"
                    }
                }
            }
        },
...
        "dao.PrivateLogisticsRecord": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.PrivateSubItem"
                    }
                }
            }
        },
        "dao.PrivateSubItem": {
            "type": "object",
            "properties": {
                "operator": {
                    "type": "string"
                },
                "peer_id": {
                    "type": "string"
                },
                "record_time": {
                    "type": "string"
                }
            }
        },
        "dao.ProductInfo": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "last_modified": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                }
            }
        },
        "dao.RecordSubItem": {
            "type": "object",
            "properties": {
                "record_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateLogisticRecordInput": {
            "type": "object",
            "properties": {
                "operator": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tracking_num": {
                    "type": "string"
                }
            }
        },
        "middleware.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errmsg": {
                    "type": "string"
                },
                "errno": {
                    "type": "integer"
                },
                "stack": {
                    "type": "object"
                },
                "trace_id": {
                    "type": "object"
                }
            }
        }
    }
}